Big commit, sorry for breaking everything.
Added a slider mode because kITE likes sliders.
Added an option for the old blending mode.
Fixed properties that used vectors not being properly animatable. This breaks old stuff.
Changed the behavior of the overlay image to work nicely by default. This breaks old stuff.
Changed the way that screen transformations work to not be resolution dependent. This breaks old stuff.